pyproject.toml
[tool.poetry]
name = "totoro-ai"
version = "0.1.0"
description = "AI engine for Totoro — intent parsing, place extraction, ranking, taste modeling"
authors = ["Saher"]
packages = [{include = "totoro_ai", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115"
uvicorn = {version = "^0.34", extras = ["standard"]}
pydantic = "^2.10"
pydantic-settings = "^2.7"
langchain = "^0.3"
langchain-openai = "^0.3"
langchain-anthropic = "^0.3"
langgraph = "^0.3"
voyageai = "^0.3"
pyyaml = "^6.0"
sqlalchemy = {version = "^2.0", extras = ["asyncio"]}
asyncpg = "^0.30"
alembic = "^1.14"
pgvector = "^0.3"
redis = {version = "^5.0", extras = ["asyncio"]}
greenlet = "^3.3.2"
instructor = "^1.0"
httpx = "^0.28"
langfuse = "^3.0"
groq = "^0.13"
yt-dlp = "^2026.3.17"
python-dotenv = "^1.0"
langgraph-checkpoint-postgres = "^3.0"
psycopg = {extras = ["binary"], version = "^3.3.3"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.25"
ruff = "^0.9"
mypy = "^1.14"
types-pyyaml = "^6.0.12.20250915"
psycopg2-binary = "^2.9.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
# LLM prompt strings, long SQL, and long error messages are inherently wide —
# line length is not enforced in files where the signal is worse than the fix.
"src/totoro_ai/core/intent/intent_parser.py" = ["E501"]
"src/totoro_ai/core/chat/router.py" = ["E501"]
"src/totoro_ai/core/chat/service.py" = ["E501"]
"src/totoro_ai/core/consult/service.py" = ["E501"]
"src/totoro_ai/core/extraction/enrichers/vision_frames.py" = ["E501"]
"src/totoro_ai/core/places/places_client.py" = ["E501"]
"src/totoro_ai/db/repositories/recall_repository.py" = ["E501"]
"src/totoro_ai/providers/llm.py" = ["E501"]
"tests/providers/test_llm_ollama.py" = ["E501"]
"src/totoro_ai/core/extraction/circuit_breaker.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "langfuse"
ignore_missing_imports = true

